requirement.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.6b0
click==8.0.1
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.8.6
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
